+++ plugin/crypto/init/init.go
@@ -1,5 +1,1 @@
 package init
-
-import (
-	_ "github.com/33cn/plugincgo/plugin/crypto/bls" //auto gen
-)