+++ plugin/dapp/init/init.go
@@ -2,10 +2,12 @@ package init
 
 import (
 	_ "github.com/33cn/plugin/plugin/dapp/autonomy"  //auto gen
+	_ "github.com/33cn/plugin/plugin/dapp/evm"       //auto gen
 	_ "github.com/33cn/plugin/plugin/dapp/hashlock"  //auto gen
 	_ "github.com/33cn/plugin/plugin/dapp/multisig"  //auto gen
 	_ "github.com/33cn/plugin/plugin/dapp/paracross" //auto gen
 	_ "github.com/33cn/plugin/plugin/dapp/retrieve"  //auto gen
+	_ "github.com/33cn/plugin/plugin/dapp/rollup"    //auto gen
 	_ "github.com/33cn/plugin/plugin/dapp/ticket"    //auto gen
 	_ "github.com/33cn/plugin/plugin/dapp/token"     //auto gen
 	_ "github.com/33cn/plugin/plugin/dapp/trade"     //auto gen